requirements.txt
# FastAPI and ASGI server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# WebRTC
aiortc==1.10.0
av==13.1.0

# Audio processing
numpy==2.2.0
soundfile==0.12.1
scipy==1.14.1
ffmpeg-python==0.2.0

# LangGraph orchestration
langgraph==0.2.59
langchain-core==0.3.28

# Environment and configuration
python-dotenv==1.0.1

# HTTP clients for external APIs
httpx==0.28.1
aiohttp==3.11.10

# OpenAI client (for OpenRouter audio API)
openai>=1.0.0

# Soundcharts official SDK
soundcharts

# Song downloader
yt-dlp==2025.12.08

# Database
aiosqlite==0.20.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2

# CORS
python-multipart==0.0.20

